backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
pydantic==2.5.0
PyJWT==2.8.0
passlib==1.7.4
aiofiles==23.2.1
python-docx==0.8.11
markdown==3.5.1
reportlab==4.0.7
aiosqlite==0.17.0